Add ax size, which watches the corpus against the split trigger
The spec settles the split in advance rather than in a hurry.
The content plane splits by year into tamnd/arxiv-20NN once a checkout goes over 20 GB or a git clone goes over fifteen minutes, whichever comes first, and the metadata, the manifests, the tags, the graph and the reports stay where they are.
This is the half of that plan that can exist before the split does, which is something that says where the corpus stands.

It walks the disk rather than asking git, because the question is what is there now and not what was committed.
work/ is left out since it is gitignored and a fresh checkout does not carry it, and .git/objects is counted apart from the checkout rather than added to it, because a clone transfers the pack and then builds the working tree.
Those are the two halves of the trigger and not one number counted twice.

The cost per paper leaves the metadata plane and the reports out of the division.
Both are in the checkout and both count against the trigger, and neither grows with the selection, so charging the metadata plane to the papers would say each of ten papers cost a gigabyte and would put the trigger a thousand times too near.

ax size -hard exits non-zero once either half has been reached, which is what makes this a watch rather than a printout.
Nothing here writes a report file, because a committed report of the checkout size changes the checkout size and would never settle.

prose.Bytes gained a gigabyte case and an int64 argument, since the two numbers it has to print are now a 500 KB figure and a 20 GB checkout.

+++ cmd/ax/figures.go
@@ -415,7 +415,7 @@ func figuresCheck(args []string) error {
 			return err
 		}
 		d := figures.Decide(*caption, im)
-		fmt.Fprintf(tw, "%s\t%s\t%d by %d\t%s\n", path, im.Format, im.Width, im.Height, prose.Bytes(im.Bytes))
+		fmt.Fprintf(tw, "%s\t%s\t%d by %d\t%s\n", path, im.Format, im.Width, im.Height, prose.Bytes(int64(im.Bytes)))
 		if im.Measured() {
 			fmt.Fprintf(tw, "  page\t%s of a page\tat %.2f by %.2f inches, from %s\n", prose.Percent(d.PageFraction), im.WidthIn, im.HeightIn, im.Stated)
 		} else {
